Implementar serviço de gerenciamento de tarefas

+++ ProcessadorTarefas/Servicos/IGerenciadorTarefas.cs
@@ -5,8 +5,8 @@ namespace ProcessadorTarefas.Servicos
     internal interface IGerenciadorTarefas
     {
         Task<Tarefa> Criar();
-        Task<Tarefa> Consultar(int idTarefa);
-        Task Cancelar(int idTarefa);
+        Task<Tarefa> Consultar(long idTarefa);
+        Task Cancelar(long idTarefa);
         Task<IEnumerable<Tarefa>> ListarAtivas();
         Task<IEnumerable<Tarefa>> ListarInativas();
     }
